+++ internal/models/refAchievementPlaces.go
@@ -5,7 +5,7 @@ import (
 )
 
 type RefAchievementPlaces struct {
-	RefAchievementPlacesID       primitive.ObjectID `bson:"_id,omitempty" json:"rewilding_places_id"`
+	RefAchievementPlacesID       primitive.ObjectID `bson:"_id,omitempty" json:"ref_achievement_places_id"`
 	RefAchievementPlacesType     string             `bson:"ref_achievement_places_type,omitempty" json:"ref_achievement_places_type,omitempty"`
 	RefAchievementPlacesName     string             `bson:"ref_achievement_places_name,omitempty" json:"ref_achievement_places_name,omitempty"`
 	RefAchievementPlacesLat      float64            `bson:"ref_achievement_places_lat,omitempty" json:"ref_achievement_places_lat,omitempty"`