+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# CHANGELOG 1.1.24
+## Changes
+- Change to achievement object
+- GET achievement/places - Retrieve 百岳, 小百岳, 保護區
+
 # CHANGELOG 1.1.23
 ## Changes
 - GET /achievement Load from rewilding event data